services/orchestrator/pyproject.toml
[project]
name = "orchestrator"
version = "0.1.0"
description = "Orchestrator service for multi-agent system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "redis>=5.2.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp>=1.28.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0",
    "opentelemetry-instrumentation-redis>=0.49b0",
    "openai-agents>=0.0.11",
    "pydantic-settings>=2.6.0",
    "structlog>=25.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"